Inserindo título, cidade, estado, salário ao objeto salário.

+++ InfojobCrawler/src/control/Job.java
@@ -5,5 +5,5 @@ public class Job {
 	String jobID;
 	String city;
 	String state;
-	int salary;
+	float salary;
 }
